Adjust types
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@732 75748be7-0387-4326-94c7-f41ce0883d2c

+++ sidplayfp/SidTune.h
@@ -47,7 +47,7 @@ class SID_EXTERN SidTune
     } LoadStatus;
 
     /// Also PSID file format limit.
-    static const uint_least16_t MAX_SONGS = 256;
+    static const unsigned int MAX_SONGS = 256;
 
     /// C64KB+LOAD+PSID
     static const uint_least32_t MAX_FILELEN = 65536+2+0x7C;
